+++ components/device_cmd/test/host/stubs/esp_err.h
@@ -1,0 +1,4 @@
+#pragma once
+typedef int esp_err_t;
+#define ESP_OK 0
+#define ESP_FAIL (-1)